+++ game/shared/basecombatweapon_shared.h
@@ -86,7 +86,7 @@ class Color;
 
 namespace vgui2
 {
-	typedef unsigned long HFont;
+	typedef uint32_t HFont;
 }
 
 // -----------------------------------------
