+++ public/basehandle.h
@@ -30,7 +30,7 @@ public:
 
 	CBaseHandle();
 	CBaseHandle( const CBaseHandle &other );
-	CBaseHandle( unsigned long value );
+	CBaseHandle( uint32_t value );
 	CBaseHandle( int iEntry, int iSerialNumber );
 
 	void Init( int iEntry, int iSerialNumber );
@@ -63,7 +63,7 @@ public:
 protected:
 	// The low NUM_SERIAL_BITS hold the index. If this value is less than MAX_EDICTS, then the entity is networkable.
 	// The high NUM_SERIAL_NUM_BITS bits are the serial number.
-	unsigned long	m_Index;
+	uint32_t	m_Index;
 };
 
 
@@ -80,7 +80,7 @@ inline CBaseHandle::CBaseHandle( const CBaseHandle &other )
 	m_Index = other.m_Index;
 }
 
-inline CBaseHandle::CBaseHandle( unsigned long value )
+inline CBaseHandle::CBaseHandle( uint32_t value )
 {
 	m_Index = value;
 }
@@ -164,7 +164,7 @@ inline bool CBaseHandle::operator <( const CBaseHandle &other ) const
 
 inline bool CBaseHandle::operator <( const IHandleEntity *pEntity ) const
 {
-	unsigned long otherIndex = (pEntity) ? pEntity->GetRefEHandle().m_Index : INVALID_EHANDLE_INDEX;
+	uint32_t otherIndex = (pEntity) ? pEntity->GetRefEHandle().m_Index : INVALID_EHANDLE_INDEX;
 	return m_Index < otherIndex;
 }
