+++ public/saverestoretypes.h
@@ -105,7 +105,7 @@ struct levellist_t
 
 struct EHandlePlaceholder_t // Engine does some of the game writing (alas, probably shouldn't), but can't see ehandle.h
 {
-	unsigned long i;
+	uint32_t i;
 };
 
 //-------------------------------------
