+++ public/steam/steamtypes.h
@@ -49,14 +49,14 @@ typedef short int16;
 typedef unsigned short uint16;
 typedef int int32;
 typedef unsigned int uint32;
-typedef long long int64;
-typedef unsigned long long uint64;
+typedef int64_t int64;
+typedef uint64_t uint64;
 #ifdef X64BITS
-typedef long long intp;
-typedef unsigned long long uintp;
+typedef int64_t intp;
+typedef uint64_t uintp;
 #else
-typedef int intp;
-typedef unsigned int uintp;
+typedef int32_t intp;
+typedef uint32_t uintp;
 #endif
 
 #endif // else _WIN32
